src/equation/gradient/mod.rs
//! Module for gradient calculation in finite element and finite volume methods.
//!
//! This module provides a flexible framework for computing gradients using
//! different numerical methods. It defines the `Gradient` struct, which serves
//! as the main interface for gradient computation, and supports multiple
//! gradient calculation methods via the `GradientCalculationMethod` enum and
//! `GradientMethod` trait.

use crate::boundary::bc_handler::BoundaryConditionHandler;
use crate::domain::{mesh::Mesh, MeshEntity, Section};
use crate::geometry::Geometry;
use std::error::Error;

pub mod gradient_calc;
pub mod tests;

use gradient_calc::FiniteVolumeGradient;

/// Enum representing the available gradient calculation methods.
pub enum GradientCalculationMethod {
    FiniteVolume,
    // Additional methods can be added here as needed
}

impl GradientCalculationMethod {
    /// Factory function to create a specific gradient calculation method based on the enum variant.
    pub fn create_method(&self) -> Box<dyn GradientMethod> {
        match self {
            GradientCalculationMethod::FiniteVolume => Box::new(FiniteVolumeGradient {}),
            // Extend here with other methods as needed
        }
    }
}

/// Trait defining the interface for gradient calculation methods.
///
/// Each gradient calculation method must implement this trait, which includes
/// the `calculate_gradient` function for computing the gradient at a given cell.
pub trait GradientMethod {
    /// Computes the gradient for a given cell.
    ///
    /// # Parameters
    /// - `mesh`: Reference to the mesh structure containing cells and faces.
    /// - `boundary_handler`: Reference to the boundary condition handler.
    /// - `geometry`: Geometry utilities for computing areas, volumes, etc.
    /// - `field`: Scalar field values for each cell.
    /// - `cell`: The current cell for which the gradient is computed.
    /// - `time`: Current simulation time.
    ///
    /// # Returns
    /// - `Ok([f64; 3])`: Computed gradient vector.
    /// - `Err(Box<dyn Error>)`: If any error occurs during computation.
    fn calculate_gradient(
        &self,
        mesh: &Mesh,
        boundary_handler: &BoundaryConditionHandler,
        geometry: &mut Geometry,
        field: &Section<f64>,
        cell: &MeshEntity,
        time: f64,
    ) -> Result<[f64; 3], Box<dyn Error>>;
}

/// Gradient calculator that accepts a gradient method for flexible computation.
///
/// This struct serves as the main interface for computing gradients across the mesh.
/// It delegates the actual gradient computation to the specified `GradientMethod`.
pub struct Gradient<'a> {
    mesh: &'a Mesh,
    boundary_handler: &'a BoundaryConditionHandler,
    geometry: Geometry,
    method: Box<dyn GradientMethod>,
}

impl<'a> Gradient<'a> {
    /// Constructs a new `Gradient` calculator with the specified calculation method.
    ///
    /// # Parameters
    /// - `mesh`: Reference to the mesh structure.
    /// - `boundary_handler`: Reference to the boundary condition handler.
    /// - `method`: The gradient calculation method to use.
    pub fn new(
        mesh: &'a Mesh,
        boundary_handler: &'a BoundaryConditionHandler,
        method: GradientCalculationMethod,
    ) -> Self {
        Self {
            mesh,
            boundary_handler,
            geometry: Geometry::new(),
            method: method.create_method(),
        }
    }

    /// Computes the gradient of a scalar field across each cell in the mesh.
    ///
    /// # Parameters
    /// - `field`: Scalar field values for each cell.
    /// - `gradient`: Mutable section to store the computed gradient vectors.
    /// - `time`: Current simulation time.
    ///
    /// # Returns
    /// - `Ok(())`: If gradients are successfully computed for all cells.
    /// - `Err(Box<dyn Error>)`: If any error occurs during computation.
    pub fn compute_gradient(
        &mut self,  // Changed to mutable reference
        field: &Section<f64>,
        gradient: &mut Section<[f64; 3]>,
        time: f64,
    ) -> Result<(), Box<dyn Error>> {
        for cell in self.mesh.get_cells() {
            let grad_phi = self.method.calculate_gradient(
                self.mesh,
                self.boundary_handler,
                &mut self.geometry,  // Now mutable
                field,
                &cell,
                time,
            )?;
            gradient.set_data(cell, grad_phi);
        }
        Ok(())
    }
}
